requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
